actualizado :rocket: * Se convierte en composer package * Separación de clases
