+++ src/Exceptions/BaseException.php
@@ -1,0 +1,9 @@
+<?php
+
+
+namespace centauroSMS\Exceptions;
+
+class BaseException extends \Exception
+{
+    // 
+}
